Change name of method getting a user record after a user update
While refactoring the Auth class, I wanted to add a method to get the
first user record of an `/accounts:lookUp` API call. I noticed that
there already was a `getUserRecordFromResponse()` method, but it was
related to responses after calls to other API endpoints.

Renaming it to `getUserRecordFromResponseAfterUserUpdate()` makes the
intent clearer.